% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgal4headers-package.R
\docType{package}
\name{cgal4headers-package}
\alias{cgal4headers}
\alias{cgal4headers-package}
\title{Provides CGAL 'C++' header-only library.}
\description{
CGAL is a C++ library that aims to provide easy access to efficient and reliable algorithms in computational geometry.
}
\examples{
\dontrun{

# Include a dependency to cgal4headers inside a cpp file
// [[Rcpp::depends(cgal4headers)]]
include <Rcpp.h>
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/dickoa/cgal4headers}
  \item Report bugs at \url{https://gitlab.com/dickoa/cgal4headers/issues}
}

}
\author{
\strong{Maintainer}: Ahmadou Dicko \email{mail@ahmadoudicko.com}

}
